Add files via upload
This is a reupload of an exploratory analysis file created with Excel and synthetic patient data related to Massachusetts General Hospital. The file name had been renamed in order to include the name of the application/tool used to create the file as well as to emphasize the fact that synthetic data had been used. After reuploading this file to the new repository dedicated to the Massachusetts General Hospital project, the previous version of this file in the "general data analytics project" repository will be deleted.